requirements.txt
Django==3.2.9
psycopg2-binary==2.9.2
djangorestframework==3.12.4
pottery==2.0.0
django-redis==5.0.0
redis==4.0.2
celery==5.2.1
tzdata==2023.3
django-celery-beat==2.2.1
aiokafka==0.7.2
kafka-python==2.0.2
supervisor==4.2.2
django-mptt==0.13.4
PyJWT==2.3.0
requests==2.27.1
drf-spectacular==0.22.1
drf-spectacular-sidecar==2022.6.1
python-keycloak==3.3.0
